Move all model logic into concerns so that models can be composed instead of inherited allowing bypassing Rails' problems with non-abstract base classes and polymorphic associations
