+++ app/models/rockauth/user.rb
@@ -1,5 +1,3 @@
-require 'active_record'
-
 module Rockauth
   class User < ActiveRecord::Base
     self.table_name = 'users'
@@ -7,28 +5,6 @@ module Rockauth
 
     resource_owner
 
-    validates_presence_of   :email, if: :email_required?
-    validates_uniqueness_of :email, case_sensitive: false, allow_nil: true
-    validates_format_of     :email, with: /\A[^@\s]+@([^@\s]+\.)+[^@\W]+\z/, if: -> { (new_record? || email_changed?) }, allow_blank: true
-
-    has_secure_password   validations: false
-    validates_presence_of :password, if: :password_required?
-    validates_length_of   :password, in: Configuration.allowed_password_length, allow_nil: true, allow_blank: true
-
-    scope :with_username, -> (username) { where("#{self.table_name}.email ILIKE ?", username) }
-
-    def self.active_model_serializer
-      Rockauth::Configuration.serializers.user.safe_constantize
-    end
-
-    delegate :active_model_serializer, to: :class
-
-    def email_required?
-      new_record? && provider_authentications.empty?
-    end
-
-    def password_required?
-      new_record? && provider_authentications.empty?
-    end
+    include Models::User
   end
 end
