+++ lib/generators/templates/provider_authentication.rb
@@ -1,2 +1,5 @@
-class ProviderAuthentication < Rockauth::ProviderAuthentication
+class ProviderAuthentication < ActiveRecord::Base
+  include Rockauth::Models::ProviderValidation
+  include Rockauth::Models::ProviderAuthentication
+  provider_authentication authentication_class_name: '::Authentication'
 end
