+++ lib/generators/templates/user.rb
@@ -1,3 +1,7 @@
-class User < Rockauth::User
+class User < ActiveRecord::Base
+  include Rockauth::Models::ResourceOwner
+
   resource_owner provider_authentication_class_name: '::ProviderAuthentication', authentication_class_name: '::Authentication'
+
+  include Rockauth::Models::User
 end
